Fix movement keys to use virtual lines
When wrapping is on, the movement keys would jump to the next real line
instead of going into the wrapped lines. This is quite annoying so I
disabled this. But then using 10j for example is not matching the
relative line numbers. To fix this the movements will use real lines
when a number is specified, otherwise a virtual line.